requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.testing.in, then run this to update:
#
#    pip-compile requirements.testing.in
#
astroid==1.4.7            # via pylint
backports-abc==0.4        # via tornado
certifi==2016.2.28        # via tornado
click==6.6
colorama==0.3.7           # via pylint
cov-core==1.13.0          # via pytest-cov
coverage==3.7.1           # via cov-core
cycler==0.10.0            # via matplotlib
docutils==0.12            # via python-daemon
lazy-object-proxy==1.2.2  # via astroid
lockfile==0.12.2          # via python-daemon
luigi==2.2.0
matplotlib==1.5.1
mock==1.0.1
numpy==1.11.1
pandas==0.18.1
pg8000==1.10.6            # via testing.postgresql
psycopg2==2.6.2
py==1.4.30                # via pytest
pylint==1.5.4
pyparsing==2.1.5          # via matplotlib
pytest-cov==1.7.0
pytest==2.8.7
python-daemon==2.1.1      # via luigi
python-dateutil==2.5.3    # via matplotlib, pandas
pytz==2016.6.1            # via matplotlib, pandas
pyyaml==3.11
scikit-learn==0.17.1
scipy==0.17.1
singledispatch==3.4.0.3   # via tornado
six==1.10.0               # via astroid, cycler, pg8000, pylint, python-dateutil, singledispatch
testing.common.database==1.1.0  # via testing.postgresql
testing.postgresql==1.3.0
tornado==4.4              # via luigi
wrapt==1.10.8             # via astroid

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via python-daemon
